requirements.txt
amqp==5.0.6
asgiref==3.4.1
billiard==3.6.4.0
celery==5.2.1
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
Django==3.2
django-extensions==3.1.5
djangorestframework==3.12.4
dramatiq==1.12.0
drf-yasg==1.20.0
gevent==21.8.0
greenlet==1.1.2
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
kombu==5.2.2
librabbitmq==2.0.0
MarkupSafe==2.0.1
msgpack==1.0.2
neovim==0.3.1
packaging==21.3
pika==1.2.0
prometheus-client==0.12.0
prompt-toolkit==3.0.22
pynvim==0.4.3
pyparsing==3.0.6
pytz==2021.3
PyYAML==6.0
requests==2.26.0
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.7
vine==5.0.0
watchdog==2.1.6
watchdog-gevent==0.1.1
wcwidth==0.2.5
Werkzeug==2.0.2
zope.event==4.5.0
zope.interface==5.4.0
